pyproject.toml
[tool.poetry]
name = "atap-llm-classifier"
version = "0.1.0"
description = ""
authors = ["Huen Chan <huen.chan@sydney.edu.au>"]
license = "MIT"
readme = "README.md"
#package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
panel = "^1.4.0"
pyyaml = "^6.0.1"
openai = "^1.13.3"
pydantic-settings = "^2.2.1"
ipykernel = "^6.29.3"
litellm = "^1.33.1"
atap-corpus-loader = ">=1.5.0"
loguru = "^0.7.2"

[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"


[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.1.5"
jupyterlab-vim = "^4.1.3"
ruff = "^0.4.3"
pytest = "^8.2.0"


[tool.poetry.scripts]
atapllmc = "atap_llm_classifier.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
